added: context

+++ app/backend/src/task/task.controller.ts
@@ -42,6 +42,7 @@ export class TaskController {
         @Request() req,
     ) {
         try {
+            if (!req.user.id) throw new BadRequestException('No hay ningun usuario logueado')
             data.userId = req.user.id
             const createdTask = await this.taskService.createTask(data)
             return createdTask
@@ -51,8 +52,11 @@ export class TaskController {
     }
     @Put(':id')
     @UseGuards(AuthGuard)
-    async updateTask(@Param('id') id: string, @Body() data: Task) {
+    async updateTask(@Param('id') id: string, @Body() data: Task,
+        @Request() req) {
         try {
+            const taskFound = await this.taskService.getTaskById(Number(id))
+            if (taskFound.userId !== req.user.id) throw new BadRequestException('No puedes actualizar una tarea que no te pertenece')
             const updatedTask = await this.taskService.updateTask(Number(id), data)
             return updatedTask
         } catch (error) {
@@ -61,8 +65,12 @@ export class TaskController {
     }
     @Put('updateState/:id')
     @UseGuards(AuthGuard)
-    async updateState(@Param('id') id: string) {
+    async updateState(
+        @Request() req,
+        @Param('id') id: string) {
         try {
+            const taskFound = await this.taskService.getTaskById(Number(id))
+            if (taskFound.userId !== req.user.id) throw new BadRequestException('No puedes actualizar una tarea que no te pertenece')
             const updatedTask = await this.taskService.updateState(Number(id))
             return updatedTask
         } catch (error) {
@@ -73,8 +81,12 @@ export class TaskController {
 
     @Delete(':id')
     @UseGuards(AuthGuard)
-    async deleteTask(@Param('id') id: string) {
+    async deleteTask(@Param('id') id: string,
+        @Request() req
+    ) {
         try {
+            const taskFound = await this.taskService.getTaskById(Number(id))
+            if (taskFound.userId !== req.user.id) throw new BadRequestException('No puedes actualizar una tarea que no te pertenece')
             return await this.taskService.deleteTask(Number(id))
         } catch (error) {
             throw new NotFoundException('La tarea no se encontró')
